refactor: Board의 Color 네이밍 변경
